+++ src/theme/breakpoints.ts
@@ -1,0 +1,10 @@
+const breakpoints = {
+  base: '0px',
+  sm: '480px',
+  md: '768px',
+  lg: '992px',
+  xl: '1300px',
+  '2xl': '1536px',
+};
+
+export default breakpoints;
